pyproject: require python-libevdev 0.13 or later
Technically we require 0.13.1 because that fixes the Python 3.9
support but let's not narrow this down too much.

python-libevdev 0.13 added a bunch of typing information which is
convenient but more importantly it adds the ability to create input
properties and event codes should they be missing in the built-in set.
This was possible before but much more awkward, now we have proper
dataclasses that can be instantiated.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Part-of: <https://gitlab.freedesktop.org/libevdev/hid-tools/-/merge_requests/180>